fix Summary, Components folder, initializing the collection
