+++ Polyclinic/Polyclinic/Appointment.cs
@@ -1,4 +1,4 @@
-﻿namespace Polyclinic.Components;
+﻿namespace Polyclinic;
 
 /// <summary>
 /// A class that describes a patient's appointment with a doctor
