--- Polyclinic/Polyclinic/Components/Enum/BloodGroup.cs
@@ -1,12 +1,0 @@
-﻿namespace Polyclinic.Components.Enum;
-
-/// <summary>
-/// The patient's blood type
-/// </summary>
-public enum BloodGroup
-{
-    I,
-    II,
-    III,
-    IV
-}
