--- Polyclinic/Polyclinic/Components/Enum/Gender.cs
@@ -1,11 +1,0 @@
-﻿namespace Polyclinic.Components.Enum;
-
-
-/// <summary>
-/// The gender of the patient
-/// </summary>
-public enum Gender
-{
-    Male,
-    Female
-}
