--- Polyclinic/Polyclinic/Components/Enum/RhesusFactor.cs
@@ -1,10 +1,0 @@
-﻿namespace Polyclinic.Components.Enum;
-
-/// <summary>
-/// The patient's Rh factor
-/// </summary>
-public enum RhesusFactor
-{
-    Positive,
-    Negative
-}
