--- Polyclinic/Polyclinic/Components/Enum/Specialization.cs
@@ -1,18 +1,0 @@
-﻿namespace Polyclinic.Components.Enum;
-
-/// <summary>
-/// Medical specialization of a doctor
-/// </summary>
-public enum Specialization
-{
-    Therapist,
-    Surgeon,
-    Cardiologist,
-    Neurologist,
-    Pediatrician,
-    Dentist,
-    Ophthalmologist,
-    Dermatologist,
-    Orthopedist,
-    Psychiatrist
-}
