+++ Polyclinic/Polyclinic/Enum/BloodGroup.cs
@@ -1,0 +1,27 @@
+﻿namespace Polyclinic.Enum;
+
+/// <summary>
+/// The patient's blood type
+/// </summary>
+public enum BloodGroup
+{
+    /// <summary>
+    /// Blood group I (O)
+    /// </summary>
+    I,
+
+    /// <summary>
+    /// Blood group II (A)
+    /// </summary>
+    II,
+
+    /// <summary>
+    /// Blood group III (B)
+    /// </summary>
+    III,
+
+    /// <summary>
+    /// Blood group IV (AB)
+    /// </summary>
+    IV
+}
