+++ Polyclinic/Polyclinic/Enum/Gender.cs
@@ -1,0 +1,17 @@
+﻿namespace Polyclinic.Enum;
+
+/// <summary>
+/// The gender of the patient
+/// </summary>
+public enum Gender
+{
+    /// <summary>
+    /// Male gender
+    /// </summary>
+    Male,
+
+    /// <summary>
+    /// Female gender
+    /// </summary>
+    Female
+}
