+++ Polyclinic/Polyclinic/Enum/RhesusFactor.cs
@@ -1,0 +1,17 @@
+﻿namespace Polyclinic.Enum;
+
+/// <summary>
+/// The patient's Rh factor
+/// </summary>
+public enum RhesusFactor
+{
+    /// <summary>
+    /// Positive Rh factor
+    /// </summary>
+    Positive,
+
+    /// <summary>
+    /// Negative Rh factor
+    /// </summary>
+    Negative
+}
