+++ Polyclinic/Polyclinic/Enum/Specialization.cs
@@ -1,0 +1,57 @@
+﻿namespace Polyclinic.Enum;
+
+/// <summary>
+/// Medical specialization of a doctor
+/// </summary>
+public enum Specialization
+{
+    /// <summary>
+    /// General practitioner specializing in internal medicine
+    /// </summary>
+    Therapist,
+
+    /// <summary>
+    /// Doctor specializing in surgical procedures
+    /// </summary>
+    Surgeon,
+
+    /// <summary>
+    /// Specialist in heart and cardiovascular system diseases
+    /// </summary>
+    Cardiologist,
+
+    /// <summary>
+    /// Specialist in nervous system disorders
+    /// </summary>
+    Neurologist,
+
+    /// <summary>
+    /// Doctor specializing in children's health
+    /// </summary>
+    Pediatrician,
+
+    /// <summary>
+    /// Specialist in oral health and dental care
+    /// </summary>
+    Dentist,
+
+    /// <summary>
+    /// Specialist in eye diseases and vision care
+    /// </summary>
+    Ophthalmologist,
+
+    /// <summary>
+    /// Specialist in skin, hair, and nail conditions
+    /// </summary>
+    Dermatologist,
+
+    /// <summary>
+    /// Specialist in musculoskeletal system disorders
+    /// </summary>
+    Orthopedist,
+
+    /// <summary>
+    /// Specialist in mental health and psychiatric disorders
+    /// </summary>
+    Psychiatrist
+}
